+++ src/js/constants.js
@@ -1,0 +1,2 @@
+const LOG_MARKER = '\u270E';
+const LOG_BADGE_DEBUG = '\u{1F7E3}';